// apps/web-naive/src/apis/models/AddDeptCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddDeptCommand = {
  parentId: number;
  deptName: string;
  orderNum: number;
  leaderName?: string;
  phone?: string;
  email?: string;
  status?: number;
};