// apps/web-naive/src/apis/models/AddRoleCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddRoleCommand = {
  roleName: string;
  roleKey: string;
  roleSort: number;
  remark?: string;
  dataScope?: string;
  status?: string;
  menuIds: Array<number>;
};