// apps/web-naive/src/apis/models/AddUserCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddUserCommand = {
  deptId?: number;
  username?: string;
  nickname?: string;
  email?: string;
  phoneNumber?: string;
  sex?: number;
  avatar?: string;
  password?: string;
  status?: number;
  roleId?: number;
  postId?: number;
  remark?: string;
};