// apps/web-naive/src/apis/models/AllocatedRoleQuery.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AllocatedRoleQuery = {
  orderColumn?: string;
  orderDirection?: string;
  timeRangeColumn?: string;
  beginTime?: string;
  endTime?: string;
  pageNum?: number;
  pageSize?: number;
  roleId?: number;
  username?: string;
  phoneNumber?: string;
};