// apps/web-naive/src/apis/models/OperationLogQuery.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OperationLogQuery = {
  orderColumn?: string;
  orderDirection?: string;
  timeRangeColumn?: string;
  beginTime?: string;
  endTime?: string;
  pageNum?: number;
  pageSize?: number;
  businessType?: string;
  status?: string;
  username?: string;
  requestModule?: string;
};