// apps/web-naive/src/apis/models/PostQuery.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PostQuery = {
  orderColumn?: string;
  orderDirection?: string;
  timeRangeColumn?: string;
  beginTime?: string;
  endTime?: string;
  pageNum?: number;
  pageSize?: number;
  postCode?: string;
  postName?: string;
  status?: number;
};