// apps/web-naive/src/apis/models/UnallocatedRoleQuery.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UnallocatedRoleQuery = {
  orderColumn?: string;
  orderDirection?: string;
  timeRangeColumn?: string;
  beginTime?: string;
  endTime?: string;
  pageNum?: number;
  pageSize?: number;
  roleId?: number;
  username?: string;
  phoneNumber?: string;
};