// apps/web-naive/src/apis/models/UpdateDeptCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateDeptCommand = {
  parentId: number;
  deptName: string;
  orderNum: number;
  leaderName?: string;
  phone?: string;
  email?: string;
  status?: number;
  deptId: number;
};